LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/abort.sub.window-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/echo' failed: A server with the specified hostname could not be found.
CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/echo' failed: A server with the specified hostname could not be found.

PASS document.open() does not abort documents that are not navigating (XMLHttpRequest)
PASS document.open() does not abort documents that are not navigating (fetch())
PASS document.open() does not abort documents that are not navigating (image loading)
FAIL document.open() does not abort documents that are not navigating (establish a WebSocket connection) assert_unreached: WebSocket should have no error Reached unreachable code
FAIL document.open() does not abort documents that are not navigating (already established WebSocket connection) assert_unreached: WebSocket should have no error Reached unreachable code
